Split Bytes Class to its own file
