+++ Makefile
@@ -1,40 +1,49 @@
-cminus: obj/lex.yy.o obj/parser.o obj/util.o obj/memory.o obj/analyze.o obj/symtab.o obj/codegenerator.o obj/syntaticErrors.o obj/instructions.o obj/library.o obj/mifgenerator.o
-	g++ -Wall -g -std=c++11 -o bin/cminus obj/*  main.cc -ly -lfl
+ANALYSIS = obj/lex.yy.o obj/parser.o obj/util.o obj/memory.o obj/analyze.o obj/symtab.o 
+SYNTHESIS = obj/codegenerator.o obj/syntaticErrors.o obj/instructions.o obj/library.o obj/mifgenerator.o
+HELPER = obj/bytes.o
+CC = g++
+FLAGS = -Wall -g -std=c++11
+
+cminus: $(ANALYSIS) $(SYNTHESIS) $(HELPER)
+	$(CC) ${FLAGS} -o bin/cminus obj/*  main.cc -ly -lfl
 
 obj/lex.yy.o: scanner.l
 	flex -o scanner.c scanner.l
-	g++  -g -std=c++11 -c scanner.c -o obj/lex.yy.o
+	$(CC)  -g -std=c++11 -c scanner.c -o obj/lex.yy.o
 
 obj/util.o: util.cc
-	g++ -g -Wall -std=c++11 -c util.cc -o obj/util.o
+	$(CC) $(FLAGS) -c util.cc -o obj/util.o
 
 obj/parser.o: cminus.y cminus.tab.c
 	bison -d cminus.y
-	g++ -g -Wall -std=c++11 -c cminus.tab.c -o obj/parser.o
+	$(CC) $(FLAGS) -c cminus.tab.c -o obj/parser.o
 
 obj/codegenerator.o: codegenerator.cc
-	g++ -g -Wall -std=c++11 -c codegenerator.cc -o obj/codegenerator.o
+	$(CC) $(FLAGS) -c codegenerator.cc -o obj/codegenerator.o
 
 obj/instructions.o: instructions.cc
-	g++ -g -Wall -std=c++11 -c instructions.cc -o obj/instructions.o
+	$(CC) $(FLAGS) -c instructions.cc -o obj/instructions.o
 
 obj/memory.o: memory.cc
-	g++ -g -Wall -std=c++11 -c memory.cc -o obj/memory.o
+	$(CC) $(FLAGS) -c memory.cc -o obj/memory.o
 
 obj/analyze.o: analyze.cc
-	g++ -g -Wall -std=c++11 -c analyze.cc -o obj/analyze.o
+	$(CC) $(FLAGS) -c analyze.cc -o obj/analyze.o
 
 obj/syntaticErrors.o: syntaticErrors.cc
-	g++ -g -Wall -std=c++11 -c syntaticErrors.cc -o obj/syntaticErrors.o
+	$(CC) $(FLAGS) -c syntaticErrors.cc -o obj/syntaticErrors.o
 		
 obj/symtab.o: symtab.cc
-	g++ -g -c -Wall -std=c++11 symtab.cc -o obj/symtab.o
+	$(CC) $(FLAGS) -c symtab.cc -o obj/symtab.o
 
 obj/library.o: library.cc
-	g++ -g -c -Wall -std=c++11 library.cc -o obj/library.o
+	$(CC) $(FLAGS) -c library.cc -o obj/library.o
 
 obj/mifgenerator.o: mifgenerator.cc
-	g++ -g -c -Wall -std=c++11 mifgenerator.cc -o obj/mifgenerator.o
+	$(CC) $(FLAGS) -c mifgenerator.cc -o obj/mifgenerator.o
+
+obj/bytes.o: bytes.cc
+	$(CC) $(FLAGS) -c bytes.cc -o obj/bytes.o
 
 view:
 	dot -Tps calc.dot -o graph.ps; evince graph.ps
