+++ codegenerator.cc
@@ -4,10 +4,10 @@
 #include "symtab.h"
 #include "memory.h"
 #include <iostream>
-#include <bitset>
 #include <assert.h>
 #include <fstream>
 #include "mifgenerator.h"
+#include "bytes.h"
 
 ConditionCodes translateCondition(TokenType operation)
 {
@@ -89,7 +89,6 @@ void CodeGenerator::linker()
     for (auto origin : labelOriginMap)
     {
         std::string label = origin.first;
-        std::cout << "LINKER label: " << label << "\n";
         for (BranchLabel *label_dest : origin.second)
         {
             int destinationAddress = labelDestMap[label]->relativeAddress;
@@ -1103,26 +1102,4 @@ std::string printRegister(int reg)
         return "$SR";
     }
     return "!UNKNOWN!";
-}
-
-Bytes::Bytes(int number)
-{
-    for (int i = 0; i < 4; i++)
-        bytes[i] = convertNthByteToInt(number, i);
-}
-
-int Bytes::getNthByte(int nthByte)
-{
-    return bytes[nthByte];
-}
-
-int Bytes::convertNthByteToInt(int number, int nthByte)
-{
-    std::string fullNumber = std::bitset<32>(number).to_string();
-    std::string byteNumber = fullNumber.substr(8 * nthByte, 8);
-    std::bitset<8> partialNumber(byteNumber);
-    int n = (int)partialNumber.to_ulong();
-    // std::cout << "This is Bytes, returning " << n << "from the number "
-            //   << number << ". That should be the " << nthByte << " position byte\n";
-    return n;
 }
